Added admin and user verification

+++ Inquirist/src/main/java/com/example/demo/controller/UserController.java
@@ -3,8 +3,11 @@ package com.example.demo.controller;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -12,6 +15,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.view.RedirectView;
 
+import com.example.demo.SecurityToolBox;
 import com.example.demo.model.Poll;
 import com.example.demo.model.User;
 import com.example.demo.model.VoteUser;
@@ -28,17 +32,21 @@ public class UserController
 
 	@Autowired
 	PollsRepository pollsRepo;
-	
+
 	@Autowired
 	VoteUsersRepository voteusersRepository;
 
 	@GetMapping
-	public ModelAndView users()
+	public String users(Map<String, Object> model)
 	{
-		// TODO only if admin
-		ModelAndView mav = new ModelAndView();
-		mav.addObject("users", usersRepo.findAll());
-		return mav;
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+
+		if (SecurityToolBox.containsRole(auth, "ROLE_ADMIN"))
+		{
+			model.put("users", usersRepo.findAll());
+		}
+
+		return "users";
 	}
 
 	@GetMapping("/{id}")
@@ -51,11 +59,16 @@ public class UserController
 		return "user_detail";
 	}
 
-	@GetMapping("/{id}/remove")
-	public RedirectView delete(@PathVariable("id") long id, Map<String, Object> model)
+	@GetMapping("{id}/remove")
+	public RedirectView deleteUser(@PathVariable("id") long id, Map<String, Object> model)
 	{
-		// TODO only if admin
-		usersRepo.deleteById(id);
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+
+		if (SecurityToolBox.containsRole(auth, "ROLE_ADMIN"))
+		{
+			usersRepo.deleteById(id);
+		}
+
 		return new RedirectView("/users");
 	}
 }
